+++ src/format_string.c
@@ -45,7 +45,7 @@ rcutils_format_string_limit(
   va_list args2;
   va_copy(args2, args1);
   // first calculate the output string
-  size_t bytes_to_be_written = rcutils_vsnprintf(NULL, 0, format_string, args1);
+  size_t bytes_to_be_written = (size_t)rcutils_vsnprintf(NULL, 0, format_string, args1);
   va_end(args1);
   // allocate space for the return string
   if (bytes_to_be_written + 1 > limit) {
